chore: Update npm dependencies and add XIcon component
The package.json file has been updated to include the latest versions of the "@nextui-org/react" and "framer-motion" dependencies. Additionally, a new component called XIcon has been added to the app/Assets/Icons directory. This component is a SVG icon for displaying an "X" symbol. These changes improve the functionality and visual appearance of the application.